# docs/reference/generated/api/openapi.json
{
  "components": {
    "schemas": {
      "AgencyResponse": {
        "additionalProperties": false,
        "description": "The GET /agencies/{agencyId} lookup body \u2014 confirms the agency exists.",
        "properties": {
          "agencyId": {
            "description": "The agency's unique id (UUID).",
            "title": "Agencyid",
            "type": "string"
          },
          "name": {
            "description": "Human-readable agency name.",
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "description": "URL-safe unique agency slug.",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "agencyId",
          "name",
          "slug"
        ],
        "title": "AgencyResponse",
        "type": "object"
      },
      "AmlRuleType": {
        "description": "The canonical kind of deterministic AML rule; the engine dispatch key (plan \u00a79.1).",
        "enum": [
          "structuring",
          "velocity",
          "high_risk_geography",
          "round_amount",
          "threshold_evasion",
          "rapid_movement"
        ],
        "title": "AmlRuleType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelVersionListResponse": {
        "additionalProperties": false,
        "description": "The model-registry versions plus the label of the currently active version.",
        "properties": {
          "activeVersionLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Label of the active (deployed) version, if any.",
            "title": "Activeversionlabel"
          },
          "versions": {
            "description": "Registry versions, newest first.",
            "items": {
              "$ref": "#/components/schemas/ModelVersionResponse"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "title": "ModelVersionListResponse",
        "type": "object"
      },
      "ModelVersionResponse": {
        "additionalProperties": false,
        "description": "One model-registry version projected onto the API surface (PHI-free).",
        "properties": {
          "artifactUri": {
            "description": "Artifact bundle uri (relative to the store).",
            "title": "Artifacturi",
            "type": "string"
          },
          "createdAt": {
            "description": "When the version was registered.",
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "featureSpec": {
            "additionalProperties": true,
            "description": "Ordered feature spec (names only).",
            "title": "Featurespec",
            "type": "object"
          },
          "metrics": {
            "additionalProperties": true,
            "description": "PHI-free holdout metrics recorded.",
            "title": "Metrics",
            "type": "object"
          },
          "notes": {
            "description": "Human-readable notes about the version.",
            "title": "Notes",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ModelVersionStatus",
            "description": "Registry lifecycle status."
          },
          "versionId": {
            "description": "The model version's unique id (UUID).",
            "title": "Versionid",
            "type": "string"
          },
          "versionLabel": {
            "description": "Unique human-readable version label.",
            "title": "Versionlabel",
            "type": "string"
          }
        },
        "required": [
          "versionId",
          "versionLabel",
          "status",
          "artifactUri",
          "featureSpec",
          "metrics",
          "notes",
          "createdAt"
        ],
        "title": "ModelVersionResponse",
        "type": "object"
      },
      "ModelVersionStatus": {
        "description": "Registry lifecycle state of a model version (plan \u00a79.2 / \u00a710.5).",
        "enum": [
          "candidate",
          "shadow",
          "canary",
          "active",
          "archived",
          "rejected"
        ],
        "title": "ModelVersionStatus",
        "type": "string"
      },
...
      "RuleCreateRequest": {
        "additionalProperties": false,
        "description": "A new agency-scoped rule; `code` + `ruleType` are immutable after creation.",
        "properties": {
          "code": {
            "description": "Stable per-agency rule code.",
            "maxLength": 64,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "description": {
            "default": "",
            "description": "Optional human-readable description.",
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "description": "Whether the rule is evaluated.",
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "description": "Rule display name.",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "description": "Rule-type-specific tunables (camelCase JSONB).",
            "title": "Params",
            "type": "object"
          },
          "ruleType": {
            "$ref": "#/components/schemas/AmlRuleType",
            "description": "Which built-in evaluator runs this rule."
          },
          "severity": {
            "$ref": "#/components/schemas/Severity",
            "default": "medium",
            "description": "Ordinal severity carried through to a hit."
          },
          "weight": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "1.0",
            "description": "Aggregation weight of the rule's contribution.",
            "title": "Weight"
          }
        },
        "required": [
          "code",
          "name",
          "ruleType"
        ],
        "title": "RuleCreateRequest",
        "type": "object"
      },
      "RuleListResponse": {
        "additionalProperties": false,
        "description": "The agency's own rules (its custom overrides; baseline rules are global).",
        "properties": {
          "rules": {
            "description": "The agency's rules, ordered by code.",
            "items": {
              "$ref": "#/components/schemas/RuleResponse"
            },
            "title": "Rules",
            "type": "array"
          }
        },
        "title": "RuleListResponse",
        "type": "object"
      },
      "RuleResponse": {
        "additionalProperties": false,
        "description": "A persisted rule row projected onto the API surface.",
        "properties": {
          "agencyId": {
            "description": "Owning tenant (agency) id.",
            "title": "Agencyid",
            "type": "string"
          },
          "code": {
            "description": "Stable per-agency rule code.",
            "title": "Code",
            "type": "string"
          },
          "createdAt": {
            "description": "When the rule was created.",
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "description": {
            "description": "Human-readable description.",
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "description": "Whether the rule is evaluated.",
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "description": "Rule display name.",
            "title": "Name",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "description": "Rule-type-specific tunables.",
            "title": "Params",
            "type": "object"
          },
          "ruleId": {
            "description": "The rule's unique id (UUID).",
            "title": "Ruleid",
            "type": "string"
          },
          "ruleType": {
            "$ref": "#/components/schemas/AmlRuleType",
            "description": "The rule's evaluator type."
          },
          "severity": {
            "$ref": "#/components/schemas/Severity",
            "description": "Ordinal severity."
          },
          "updatedAt": {
            "description": "When the rule was last updated.",
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "version": {
            "description": "Monotonic version (bumped on each update).",
            "title": "Version",
            "type": "integer"
          },
          "weight": {
            "description": "Aggregation weight.",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight",
            "type": "string"
          }
        },
        "required": [
          "ruleId",
          "agencyId",
          "code",
          "name",
          "description",
          "ruleType",
          "params",
          "severity",
          "weight",
          "enabled",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "title": "RuleResponse",
        "type": "object"
      },
      "RuleUpdateRequest": {
        "additionalProperties": false,
        "description": "A partial update of a rule's mutable fields; only sent fields are applied (PATCH).",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New description.",
            "title": "Description"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Enable (true) or disable (false).",
            "title": "Enabled"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New display name.",
            "title": "Name"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Replacement tunables object.",
            "title": "Params"
          },
          "severity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Severity"
              },
              {
                "type": "null"
              }
            ],
            "description": "New severity."
          },
          "weight": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "New aggregation weight.",
            "title": "Weight"
          }
        },
        "title": "RuleUpdateRequest",
        "type": "object"
      },
      "Severity": {
        "description": "Ordinal severity shared by AML rules, alerts, and drift reports.",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "FraudLens AML fraud investigation API (walking skeleton).",
    "title": "FraudLens",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agencies/{agency_id}": {
      "get": {
        "description": "Return the requested agency once tenancy is validated; 404 if it does not exist.",
        "operationId": "read_agency_api_v1_agencies__agency_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agency_id",
            "required": true,
            "schema": {
              "title": "Agency Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgencyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Agency",
        "tags": [
          "tenancy"
        ]
      }
    },
...
    "/api/v1/model-versions": {
      "get": {
        "description": "Return the model registry (newest first) plus the active version label.",
        "operationId": "list_model_versions_api_v1_model_versions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Model Versions",
        "tags": [
          "model-versions"
        ]
      }
    },
    "/api/v1/model-versions/{version_id}": {
      "get": {
        "description": "Return one model-registry version by id; 404 when it does not exist.",
        "operationId": "get_model_version_api_v1_model_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Model Version",
        "tags": [
          "model-versions"
        ]
      }
    },
    "/api/v1/rules": {
      "get": {
        "description": "Return the agency's own rules (its custom overrides), ordered by code.",
        "operationId": "list_rules_api_v1_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Rules",
        "tags": [
          "rules"
        ]
      },
      "post": {
        "description": "Create an agency-scoped rule (201); 409 when its code already exists for the agency.",
        "operationId": "create_rule_api_v1_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Rule",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/v1/rules/{rule_id}": {
      "delete": {
        "description": "Delete the agency's rule (204); 404 when missing, global, or cross-tenant.",
        "operationId": "delete_rule_api_v1_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Rule",
        "tags": [
          "rules"
        ]
      },
      "get": {
        "description": "Return one rule by id; 404 when missing, global, or owned by another agency.",
        "operationId": "get_rule_api_v1_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Rule",
        "tags": [
          "rules"
        ]
      },
      "patch": {
        "description": "Apply a partial update (incl. enable/disable) to the agency's rule; bumps version.",
        "operationId": "update_rule_api_v1_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Rule",
        "tags": [
          "rules"
        ]
      }
    },
...
  }
}